Move isMovable method from attributes classes to entities classes.
