+++ src/main/java/org/ml_methods_group/algorithm/entity/MethodEntity.java
@@ -20,10 +20,12 @@ import com.intellij.psi.PsiMethod;
 import com.sixrr.metrics.MetricCategory;
 import com.sixrr.metrics.utils.MethodUtils;
 import org.jetbrains.annotations.NotNull;
+import org.ml_methods_group.utils.PSIUtil;
 
 public class MethodEntity extends CodeEntity {
-    private final @NotNull
-    PsiMethod psiMethod;
+    private final @NotNull PsiMethod psiMethod;
+
+    private final boolean isMovable;
 
     public MethodEntity(
         final @NotNull PsiMethod psiMethod,
@@ -31,11 +33,19 @@ public class MethodEntity extends CodeEntity {
     ) {
         super(relevantProperties);
         this.psiMethod = psiMethod;
+
+        isMovable = !PSIUtil.isOverriding(psiMethod) &&
+                !MethodUtils.isAbstract(psiMethod) && !psiMethod.isConstructor();
     }
 
     @Override
     public @NotNull String getIdentifier() {
         return MethodUtils.calculateSignature(psiMethod);
+    }
+
+    @Override
+    public boolean isMovable() {
+        return isMovable;
     }
 
     @Override
